C program that prints the size of various types on the computer it is compiled and run on

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ long int admission_number;
 long long int p_number;
 float mass;
 
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(grade));
-printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(age));
-printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(admission_number));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(p_number));
-printf("Size of a float: %lu byte(s)\n" (unsigned long)sizeof(mass));
+printf("Size of a char: %zu byte(s)\n", sizeof(grade));
+printf("Size of an int: %zu byte(s)\n", sizeof(age));
+printf("Size of a long int: %zu byte(s)\n", sizeof(admission_number));
+printf("Size of a long long int: %zu byte(s)\n", sizeof(p_number));
+printf("Size of a float: %zu byte(s)\n" sizeof(mass));
 return (0);
 }
